be explicity when you call terms.data.frame
Here is a fix for issue https://github.com/tidymodels/hardhat/issues/155
If you are assuming that terms on a data.frame always calls stats::terms.default while 
- nowhere in the documentation of ?stats::terms it is indicated that you can call it on a data.frame without using terms.default 
- nor is terms.data.frame implemented in package stats
It might be better that you make your assumptions explicit. So that you don't claim terms.data.frame for other packages.
This fixes the issue for user @LasWin
All the best.