Allow casting from `<frequency_weights>` -> `<double>`
This makes sense because it is never a lossy cast. Going from `<importance_weights>` -> `<integer>` has more of a chance to be lossy, so I don't feel comfortable adding that (if you REALLY need it you can go through double first).

This also allows users of weights to be able to call `as.double()` on the weights no matter what type they are to begin with, which is nice for consistency