Restructure clean/process internals to use S3 dispatch (#201)
* Restructure clean/process internals to use S3 dispatch

We don't actually need to store the clean/process functions that support `mold()` and `forge()` in the blueprint. We can instead look up their current versions at runtime using S3 dispatch - this is what `run_mold()` and `run_forge()` are now intended to do. This means that we should never be calling an outdated clean/process function from `forge()`, and should result in less breakage.

* Update pkgdown reference

* NEWS bullets